new silence mode, voltage threshold are floats, new default for "Telemetry Check Interval"
- Silence Mode: 2-pos switch in Down-Position during power on
Transmitter turns off buzzer and uses LED instead.
- config.h accepts now directly float values for voltage thresholds
g_Profile[].V_A1[], g_Profile[].V_A2[] and g_V_TX (no calculation
needed)
- default for g_Telemetry_Check_Interval is 8000ms

+++ T5x/Tx5Types.h
@@ -12,8 +12,8 @@ typedef struct
     uint8_t       EleDR[3];
     uint8_t       RudDR[3];
 
-    uint8_t       V_A1[2];    // The internal sensor has a 4:1 divider, so the value is 0-13,2V in 255 steps or 0,052V per step. 0...warning, 1...emergency
-    uint8_t       V_A2[2];    // A2 without divider is 0-3,3V in 255 steps or 0,013V per step.                                   0...warning, 1...emergency
+    float         V_A1[2];    // The internal sensor has a 4:1 divider, so the value is 0-13,2V in 255 steps or 0,052V per step. 0...warning, 1...emergency
+    float         V_A2[2];    // A2 without divider is 0-3,3V in 255 steps or 0,013V per step.                                   0...warning, 1...emergency
     uint16_t      Timer;      // flight timer     
 } Profile_t;
